Feature/show withdraw fee (#43)
* move shared functionality to hooks

* didnt move all shared logic

* display fee under balance

* move withdrawal fee to under amount

* move fee back to under balance
