+++ web/src/components/DepositCard/DepositCard.tsx
@@ -3,8 +3,8 @@ import { useEffect, useMemo, useState } from "react";
 
 import { Dec, DecUtils } from "@keplr-wallet/unit";
 import AnimatedArrowSpacer from "components/AnimatedDownArrowSpacer/AnimatedDownArrowSpacer";
-import Dropdown, { type DropdownOption } from "components/Dropdown/Dropdown";
-import { useConfig, type EvmChainInfo, type IbcChainInfo } from "config";
+import Dropdown from "components/Dropdown/Dropdown";
+import { useConfig } from "config";
 import { useEvmChainSelection } from "features/EthWallet";
 import { sendIbcTransfer, useIbcChainSelection } from "features/KeplrWallet";
 import { useNotifications, NotificationType } from "features/Notifications";
@@ -18,55 +18,29 @@ export default function DepositCard(): React.ReactElement {
     selectEvmChain,
     evmChainsOptions,
     selectedEvmChain,
+    selectedEvmChainOption,
+    defaultEvmCurrencyOption,
     selectEvmCurrency,
     evmCurrencyOptions,
     evmBalance,
     isLoadingEvmBalance,
     connectEVMWallet,
   } = useEvmChainSelection(evmChains);
-  const defaultEvmCurrencyOption = useMemo(() => {
-    return evmCurrencyOptions[0] || null;
-  }, [evmCurrencyOptions]);
 
   const {
     ibcAccountAddress: fromAddress,
     selectIbcChain,
     ibcChainsOptions,
     selectedIbcChain,
+    selectedIbcChainOption,
+    defaultIbcCurrencyOption,
     selectIbcCurrency,
-    ibcCurrencyOptions,
     selectedIbcCurrency,
+    ibcCurrencyOptions,
     ibcBalance,
     isLoadingIbcBalance,
     connectKeplrWallet,
   } = useIbcChainSelection(ibcChains);
-  const defaultIbcCurrencyOption = useMemo(() => {
-    return ibcCurrencyOptions[0] || null;
-  }, [ibcCurrencyOptions]);
-
-  // selectedIbcChainOption allows us to ensure the label is set properly
-  // in the dropdown when connecting via an "additional option"s action,
-  //  e.g. the "Connect Keplr Wallet" option in the dropdown
-  const selectedIbcChainOption = useMemo(() => {
-    if (!selectedIbcChain) {
-      return null;
-    }
-    return {
-      label: selectedIbcChain?.chainName || "",
-      value: selectedIbcChain,
-      leftIconClass: selectedIbcChain?.iconClass || "",
-    } as DropdownOption<IbcChainInfo>;
-  }, [selectedIbcChain]);
-  const selectedEvmChainOption = useMemo(() => {
-    if (!selectedEvmChain) {
-      return null;
-    }
-    return {
-      label: selectedEvmChain?.chainName || "",
-      value: selectedEvmChain,
-      leftIconClass: selectedEvmChain?.iconClass || "",
-    } as DropdownOption<EvmChainInfo>;
-  }, [selectedEvmChain]);
 
   // the evm currency selection is controlled by the sender's chosen ibc currency,
   // and should be updated when an ibc currency or evm chain is selected
