+++ web/src/features/KeplrWallet/hooks/useIbcChainSelection.tsx
@@ -115,6 +115,24 @@ export function useIbcChainSelection(ibcChains: IbcChains) {
     );
   }, [selectedIbcChain]);
 
+  const defaultIbcCurrencyOption = useMemo(() => {
+    return ibcCurrencyOptions[0] || null;
+  }, [ibcCurrencyOptions]);
+
+  // selectedIbcChainOption allows us to ensure the label is set properly
+  // in the dropdown when connecting via an "additional option"s action,
+  //  e.g. the "Connect Keplr Wallet" option in the dropdown
+  const selectedIbcChainOption = useMemo(() => {
+    if (!selectedIbcChain) {
+      return null;
+    }
+    return {
+      label: selectedIbcChain?.chainName || "",
+      value: selectedIbcChain,
+      leftIconClass: selectedIbcChain?.iconClass || "",
+    } as DropdownOption<IbcChainInfo>;
+  }, [selectedIbcChain]);
+
   const selectIbcCurrency = useCallback((currency: IbcCurrency) => {
     setSelectedIbcCurrency(currency);
   }, []);
@@ -192,6 +210,8 @@ export function useIbcChainSelection(ibcChains: IbcChains) {
 
     selectedIbcChain,
     selectedIbcCurrency,
+    defaultIbcCurrencyOption,
+    selectedIbcChainOption,
 
     ibcAccountAddress,
     ibcBalance,
